test_project/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
mysqlclient==2.1.0
sqlparse==0.4.2
djangorestframework==3.13.1
pytz==2022.1
django-session-timeout==0.1.0
six==1.16.0
